Linux 2.10 Open Source Gold Release
Provided a reproducible SDK.
Supported new OS: RHEL 8.1, CentOS 8.1 and Fedora 31.
Supported user to specify platform id in PCK Cert ID Retrieval Tool’s command line option.
Added ability to execute Platform Cert ID Retrieval Tool on multi-package platforms without
  loading enclaves. PCCS now supports this functionality. The platform still needs to support
  SGX.
Updated Platform Cert ID Retrieval Tool and Multi-package registration tool to align with BIOS
  platform manifest changes.
Added .deb and .rpm installers for Platform Cert ID Retrieval Tool and Multi-package
  Registration Agent.
Fixed bugs.

+++ psw/ae/aesm_service/source/utils/type_length_value.cpp
@@ -171,7 +171,7 @@ static tlv_status_t tlv_msg_init_one_tlv(tlv_info_t* infos, const tlv_msg_t& tlv
     if(infos->header_size == UNKNOWN_TLV_HEADER_SIZE)
         hsize = get_tlv_header_size_from_payload_size(infos->size); //TLV header size
     else
-        hsize = (uint8_t)infos->header_size;
+        hsize = infos->header_size;
     uint32_t tsize = hsize+infos->size; //size of header and payload
     if(tlv_msg.msg_size<tsize)return TLV_INSUFFICIENT_MEMORY;
     if((status=write_tlv_header(tlv_msg.msg_buf , infos))!=TLV_SUCCESS)
